src/requirements.txt
requests
pyyaml
gitpython
boto3
tqdm
jsonschema
botocore
python-dateutil
paramiko
git-url-parse
rich
fastapi
starlette
uvicorn
pydantic